test: give the fake instance a file of its own
cmd_test.go was two thirds fake: a stub of the whole Flagsmith API, now
shared with the scripts, sitting in the file named after the tests that used
to be its only caller. It moves to fake_test.go, leaving cmd_test.go as the
fourteen tests a script cannot express plus the harness they run through.

newFake registered fifty-odd routes in one function. It now calls a mount
per resource, so the handlers for a command's endpoints sit together and the
largest of them is a screen or two rather than a thousand lines. The
authorization check the handlers shared was a closure inside it, and is now
a function beside them.

beep boop
